goblins: Set image dimensions
This PR adds `width` and `height` attributes to the goblin thumbnails.
This prevents page reflows while the images load. This required changing
the gallery CSS to use `width` instead of `max-width`, but I couldn't
see any places where that broke existing layouts, so I'm pretty sure
it's a safe change.

Also switched the `src` for Björn and Elín to use the thumbnail image
rather than the full-size one.